+++ packages/lib/plugin.js
@@ -3,9 +3,6 @@
  * @module lib/plugin
  */
 "use strict";
-const path = require("path");
-
-const libErrors = require("@clusterio/lib/errors");
 
 
 /**
@@ -387,45 +384,6 @@ class BaseControlPlugin {
 }
 
 /**
- * Load plugin information
- *
- * Loads plugin info modules for the paths to the given plugins.  Once
- * loaded the info modules will not be reloaded should this function be
- * called again.
- *
- * @param {Map<string, string>} pluginList -
- *     Mapping of plugin name to require path for the plugins to load.
- * @returns {Array<Object>} Array of plugin info modules.
- * @static
- */
-async function loadPluginInfos(pluginList) {
-	let plugins = [];
-	for (let [pluginName, pluginPath] of pluginList) {
-		let pluginInfo;
-		let pluginPackage;
-
-		try {
-			pluginInfo = require(path.posix.join(pluginPath, "info"));
-			pluginPackage = require(path.posix.join(pluginPath, "package.json"));
-
-		} catch (err) {
-			throw new libErrors.PluginError(pluginName, err);
-		}
-
-		if (pluginInfo.name !== pluginName) {
-			throw new libErrors.EnvironmentError(
-				`Expected plugin at ${pluginPath} to be named ${pluginName} but got ${pluginInfo.name}`
-			);
-		}
-
-		pluginInfo.requirePath = pluginPath;
-		pluginInfo.version = pluginPackage.version;
-		plugins.push(pluginInfo);
-	}
-	return plugins;
-}
-
-/**
  * Attach plugin messages
  *
  * Attaches all messages defined in the `.message` property of the
@@ -485,7 +443,6 @@ module.exports = {
 	BaseMasterPlugin,
 	BaseControlPlugin,
 
-	loadPluginInfos,
 	attachPluginMessages,
 	invokeHook,
 };
